added stopengine.py

+++ stopengine.py
@@ -2,5 +2,10 @@
 import os
 import subprocess
 n=subprocess.check_output("sudo lsof -t -i:8888",shell=True)
-os.system("sudo kill -9 %s" %n)
+
+if n:
+	os.system("sudo kill -9 %s" %n)
+	print("LCC service has been successfully stopped.")
+else:
+	print("Failed to find LCC service on port 8888")
